test(service-provider-server): remove duplication
